ViewHooks: touchup on constructor calls and add Dynamic view hooks subscriber
