+++ core/src/Kokkos_ViewHolder.hpp
@@ -119,7 +119,14 @@ class ConstViewHolderBase {
 
   virtual void deep_copy_to_buffer(unsigned char *buff) = 0;
 
+ protected:
+
+  explicit ConstViewHolderBase( ViewMetadata &&md )
+    : m_metadata( std::move( md ) )
+  {}
+
  private:
+
   ViewMetadata m_metadata;
 };
 
@@ -130,6 +137,12 @@ class ViewHolderBase : public ConstViewHolderBase {
   // virtual void *data() = 0;
   // virtual ViewHolderBase *clone() const = 0;
   virtual void deep_copy_from_buffer(unsigned char *buff) = 0;
+
+ protected:
+
+  explicit ViewHolderBase( ViewMetadata &&md )
+    : ConstViewHolderBase( std::move( md ) )
+  {}
 };
 
 template <typename View, typename Enable = void>
@@ -138,9 +151,9 @@ class ViewHolder : public ViewHolderBase {
   virtual ~ViewHolder() = default;
 
   explicit ViewHolder(const View &view)
-      : m_metadata(m_view.span(), m_view.span_is_contiguous(), m_view.data(),
+      : ViewHolderBase(ViewMetadata{m_view.span(), m_view.span_is_contiguous(), m_view.data(),
                    m_view.label(), sizeof(typename View::value_type),
-                   std::is_same<typename View::memory_space, HostSpace>::value),
+                   std::is_same<typename View::memory_space, HostSpace>::value}),
         m_view(view) {}
 
   void deep_copy_to_buffer(unsigned char *buff) override {
@@ -165,9 +178,9 @@ class ViewHolder<View, typename std::enable_if<std::is_const<
   virtual ~ViewHolder() = default;
 
   explicit ViewHolder(const View &view)
-      : m_metadata(m_view.span(), m_view.span_is_contiguous(), m_view.data(),
+      : ConstViewHolderBase(ViewMetadata{m_view.span(), m_view.span_is_contiguous(), m_view.data(),
                    m_view.label(), sizeof(typename View::value_type),
-                   std::is_same<typename View::memory_space, HostSpace>::value),
+                   std::is_same<typename View::memory_space, HostSpace>::value}),
         m_view(view) {}
 
   void deep_copy_to_buffer(unsigned char *buff) override {
@@ -178,76 +191,6 @@ class ViewHolder<View, typename std::enable_if<std::is_const<
  private:
   View m_view;
 };
-
-struct ViewHooks {
-  using callback_type       = std::function<void(ViewHolderBase &)>;
-  using const_callback_type = std::function<void(ConstViewHolderBase &)>;
-
-  template <typename F, typename ConstF>
-  static void set(F &&fun, ConstF &&const_fun) {
-    s_callback       = std::forward<F>(fun);
-    s_const_callback = std::forward<ConstF>(const_fun);
-  }
-
-  static void clear() {
-    s_callback       = callback_type{};
-    s_const_callback = const_callback_type{};
-  }
-
-  static bool is_set() noexcept {
-    return static_cast<bool>(s_callback) || static_cast<bool>(s_const_callback);
-  }
-
-  template <class DataType, class... Properties>
-  static void call(View<DataType, Properties...> &view) {
-    callback_type tmp_callback;
-    const_callback_type tmp_const_callback;
-
-    std::swap(s_callback, tmp_callback);
-    std::swap(s_const_callback, tmp_const_callback);
-
-    auto holder = ViewHolder<View<DataType, Properties...> >(view);
-
-    do_call(tmp_callback, tmp_const_callback, std::move(holder));
-
-    std::swap(s_callback, tmp_callback);
-    std::swap(s_const_callback, tmp_const_callback);
-  }
-
- private:
-  static void do_call(callback_type _cb, const_callback_type _ccb,
-                      ViewHolderBase &&view) {
-    if (_cb) {
-      _cb(view);
-    }
-  }
-
-  static void do_call(callback_type _cb, const_callback_type _ccb,
-                      ConstViewHolderBase &&view) {
-    if (_ccb) _ccb(view);
-  }
-
-  static callback_type s_callback;
-  static const_callback_type s_const_callback;
-};
-
-namespace Impl {
-template <class ViewType, class Traits = typename ViewType::traits,
-          class Enabled = void>
-struct ViewHooksCaller {
-  static void call(ViewType &view) {}
-};
-
-template <class ViewType, class Traits>
-struct ViewHooksCaller<
-    ViewType, Traits,
-    typename std::enable_if<!std::is_same<typename Traits::memory_space,
-                                          AnonymousSpace>::value>::type> {
-  static void call(ViewType &view) {
-    if (ViewHooks::is_set()) ViewHooks::call(view);
-  }
-};
-}  // namespace Impl
 
 }  // namespace Kokkos
